+++ pyanalyze/error_code.py
@@ -100,6 +100,7 @@ class ErrorCode(enum.Enum):
     unused_assignment = 80
     incompatible_yield = 81
     invalid_import = 82
+    too_many_positional_args = 83
 
 
 # Allow testing unannotated functions without too much fuss
@@ -120,6 +121,7 @@ DISABLED_BY_DEFAULT = {
     ErrorCode.unused_ignore,
     ErrorCode.missing_f,
     ErrorCode.bare_ignore,
+    ErrorCode.too_many_positional_args,
     # TODO: turn this on
     ErrorCode.implicit_reexport,
     ErrorCode.incompatible_override,
@@ -221,6 +223,9 @@ ERROR_DESCRIPTION = {
     ErrorCode.unused_assignment: "Assigned value is never used",
     ErrorCode.incompatible_yield: "Incompatible yield type",
     ErrorCode.invalid_import: "Invalid import",
+    ErrorCode.too_many_positional_args: (
+        "Call with many positional arguments should use keyword arguments"
+    ),
 }