+++ docs/changelog.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 - Simplify import handling; stop trying to import modules at type checking time (#566)
+- Suggest using keyword arguments on calls with too many positional arguments (#572)
+- Emit an error for unknown `TypedDict` keys (#567)
+- Improve type inference for f-strings containing literals (#571)
+- Add experimental `@has_extra_keys` decorator for `TypedDict` types (#568)
 - Fix crash on recursive type aliases. Recursive type aliases now fall back to `Any` (#565)
 - Support `in` on objects with only `__getitem__` (#564)
 - Add support for `except*` (PEP 654) (#562)